Added logout; fixes to signin; url changes
I've added a way to log out (Tracks/logout) and made some changes to
sign in to make sure that users are actually authenticated. As of now,
edits are needed to the template file to accept an optional message to
allow for error handling. I've also made all URLs terminal for now (i.e.
'aboutaboat' no longer resolves to 'about'), but this can easily be
reversed if it should be.